requirements.txt
# requirements.txt
streamlit
google-cloud-firestore
gsheetsdb==0.1.13.1